moved the search function back into the pageTurn file
