+++ src/js/handlers/listners/pageTurn.js
@@ -1,7 +1,7 @@
 import { get } from "../../api/requests/get.js";
 import { renderListings } from "../events/renderListings.js";
 import { load } from "../../storage/load.js";
-import { updateCurrentPageDisplay, updatePaginationBtns } from "../events/pagination.js";
+import { updateCurrentPageDisplay, updatePaginationBtns, search } from "../events/_index.js";
 
 const feed = document.getElementById("feed");
 let page = 1;
@@ -57,5 +57,14 @@ export function listenForPageTurn(nxtbtn, prvbtn) {
       }
     }
     window.scrollTo(0, 0);
+  });
+}
+
+export async function listenForSearch() {
+  const searchInput = document.getElementById("searchbar");
+  searchInput.addEventListener("input", async () => {
+    page = 1;
+    updateCurrentPageDisplay(page);
+    search(page);
   });
 }
